New var `outshine-org-style-global-cycling-at-bob-p' introduced.
When this is set (customized) to non-nil, TAB in outshine behaves like
in Org-mode wrt global buffer-cycling when point is at
beginning-of-buffer. Otherwise, the default behaviour, inherited (and
then modified) from outline-magic.el, is active: cycle whole buffer with
TAB, when point is at beginning of buffer and on headline.