First version of `outshine-use-outorg' added.
Since Org-mode and Outshine are just two views on the same file, and
Outorg allows to switch quickly between the two views, this function is
a first attempt to make the full Org-mode functionality available in
Outshine source-code buffers without reimplementing Org-mode.

See the Outshine speed commands for usage examples.