Fix error when calling `org-overview`
When `org-overview` or similar commands were called which then call
`outline-hide-sublevels`, the command wasn't getting executed because
it gave an error saying "Before First Headline". The issue was seen if
the org file had prelude code as below

      #+STARTUP: indent
      * First headline

The error was issued when calling the `hide-subtree` function inside
`outline-hide-sublevels`. The issue is resolved by this fix.